Removed unecessary extra wrapper component
